firmware: move max283x driver instance to max283x.x
