+++ firmware/common/drivers.h
@@ -27,7 +27,6 @@ extern "C" {
 
 #include "cpld_jtag.h"
 #include "i2c_lpc.h"
-#include "max283x.h"
 #include "max5864.h"
 #include "mixer.h"
 #include "platform_detect.h" // IWYU pragma: keep
@@ -43,7 +42,6 @@ extern "C" {
 /* TODO: Hide these configurations */
 extern const i2c_lpc_config_t i2c_config_si5351c_fast_clock;
 extern si5351c_driver_t clock_gen;
-extern ssp_config_t ssp_config_max283x;
 extern ssp_config_t ssp_config_max5864;
 
 #if defined(IS_PRALINE)
@@ -51,7 +49,6 @@ extern ssp_config_t ssp_config_ice40_fpga;
 extern ice40_spi_driver_t ice40;
 extern fpga_driver_t fpga;
 #endif
-extern max283x_driver_t max283x;
 extern max5864_driver_t max5864;
 extern mixer_driver_t mixer;
 extern sgpio_config_t sgpio_config;
@@ -59,7 +56,6 @@ extern rf_path_t rf_path;
 extern jtag_gpio_t jtag_gpio_cpld;
 extern jtag_t jtag_cpld;
 
-void ssp1_set_mode_max283x(void);
 void ssp1_set_mode_max5864(void);
 #if defined(IS_PRALINE)
 void ssp1_set_mode_ice40(void);
