+++ firmware/common/fpga_image.c
@@ -27,6 +27,7 @@
 #include "fpga.h"
 #include "ice40_spi.h"
 #include "lz4_blk.h"
+#include "max283x.h"
 #include "selftest.h"
 
 struct fpga_image_read_ctx {
