+++ firmware/common/pins.c
@@ -29,6 +29,7 @@
 #include "drivers.h"
 #include "gpio.h"
 #include "leds.h"
+#include "max283x.h"
 #include "mixer.h"
 #include "platform_detect.h"
 #include "platform_gpio.h"
